+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/ui/channels/ChannelItemTest.kt
@@ -18,7 +18,7 @@ package io.getstream.chat.android.compose.ui.channels
 
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
-import io.getstream.chat.android.compose.ui.SnapshotTest
+import io.getstream.chat.android.compose.ui.PaparazziComposeTest
 import io.getstream.chat.android.compose.ui.channels.list.ChannelItemDraftMessage
 import io.getstream.chat.android.compose.ui.channels.list.ChannelItemLastMessageSeenStatus
 import io.getstream.chat.android.compose.ui.channels.list.ChannelItemLastMessageSentStatus
@@ -28,7 +28,7 @@ import io.getstream.chat.android.compose.ui.channels.list.ChannelItemUnreadMessa
 import org.junit.Rule
 import org.junit.Test
 
-internal class ChannelItemTest : SnapshotTest {
+internal class ChannelItemTest : PaparazziComposeTest {
 
     @get:Rule
     override val paparazzi = Paparazzi(deviceConfig = DeviceConfig.PIXEL_2)