+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/ui/components/TranslatedLabelTest.kt
@@ -20,11 +20,11 @@ import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.ui.Modifier
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
-import io.getstream.chat.android.compose.ui.SnapshotTest
+import io.getstream.chat.android.compose.ui.PaparazziComposeTest
 import org.junit.Rule
 import org.junit.Test
 
-internal class TranslatedLabelTest : SnapshotTest {
+internal class TranslatedLabelTest : PaparazziComposeTest {
 
     @get:Rule
     override val paparazzi = Paparazzi(deviceConfig = DeviceConfig.PIXEL_2)