+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/ui/pinned/PinnedMessageListTest.kt
@@ -20,14 +20,14 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
-import io.getstream.chat.android.compose.ui.SnapshotTest
+import io.getstream.chat.android.compose.ui.PaparazziComposeTest
 import io.getstream.chat.android.previewdata.PreviewPinnedMessageData
 import io.getstream.chat.android.ui.common.model.MessageResult
 import io.getstream.chat.android.ui.common.state.pinned.PinnedMessageListState
 import org.junit.Rule
 import org.junit.Test
 
-internal class PinnedMessageListTest : SnapshotTest {
+internal class PinnedMessageListTest : PaparazziComposeTest {
 
     @get:Rule
     override val paparazzi = Paparazzi(deviceConfig = DeviceConfig.PIXEL_2)