+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Fix `Message.channelInfo` not populated when parsing `message.new` event. [#5953
 
 ## stream-chat-android-state
 ### 🐞 Fixed
+- Fix increasing unread count for muted channels. [#5947](https://github.com/GetStream/stream-chat-android/pull/5947)
 
 ### ⬆️ Improved